refactor: 상속을 통한 공통한 작업 추가
